Add files via upload
This notebook performs an in-depth exploratory data analysis (EDA) of the IIT Madras Punjabi text-to-speech dataset. The aim is to understand the structure and distribution of audio and textual data, examine speaker and utterance mappings, and evaluate the overall quality and consistency of the dataset for further modeling and application in a Punjabi text-to-speech synthesis project.