Add files via upload
The IIT Madras Punjabi dataset consists of speech audio recordings in Punjabi, accompanied by corresponding transcriptions. It includes a variety of speakers, both male and female, to capture diverse speech patterns. The dataset is structured with audio files in WAV format and associated metadata containing speaker information and transcriptions. This makes it suitable for speech recognition and text-to-speech model development.